Get GApplication out of experimental settings
GApplication will always be used on Linux now. GNotification gets a toggle instead.
