+++ Telegram/SourceFiles/window/notifications_manager.h
@@ -74,7 +74,19 @@ struct custom_is_fast_copy_type<Window::Notifications::ChangeType> : std::true_t
 
 } // namespace base
 
+namespace base::options {
+
+template <typename Type>
+class option;
+
+using toggle = option<bool>;
+
+} // namespace base::options
+
 namespace Window::Notifications {
+
+extern const char kOptionGNotification[];
+extern base::options::toggle OptionGNotification;
 
 class Manager;
